add token

+++ src/main/java/com/dkc/controllers/profile_officerController.java
@@ -18,8 +18,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.dkc.models.entities.profile_officer;
+import com.dkc.services.authService;
 import com.dkc.services.profile_officerService;
 
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 
 @RestController
@@ -28,11 +30,21 @@ public class profile_officerController {
 
     @Autowired 
     private profile_officerService Profile_officerService;
+    @Autowired 
+    private authService AuthService; 
 
     @PostMapping()
-    public ResponseEntity<Object> create(@Valid @RequestBody profile_officer data, BindingResult bindingResult)
+    public ResponseEntity<Object> create(@Valid @RequestBody profile_officer data, BindingResult bindingResult, HttpServletRequest request)
     {
         Map<String, Object> responseMap = new HashMap<>();
+        String token = request.getHeader("token");
+        boolean checkValid = AuthService.checkValid(token);
+        if(!checkValid) {
+            responseMap.put("message", "Not authorize!");
+            responseMap.put("data", checkValid);
+            return new ResponseEntity<>(responseMap, HttpStatus.OK);
+        }
+
         if(bindingResult.hasErrors())
         {
             responseMap.put("message", "All field required!");
@@ -83,9 +95,17 @@ public class profile_officerController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Object> update(@PathVariable("id") String id, @Valid @RequestBody profile_officer data, BindingResult bindingResult)
+    public ResponseEntity<Object> update(@PathVariable("id") String id, @Valid @RequestBody profile_officer data, BindingResult bindingResult, HttpServletRequest request)
     {
         Map<String, Object> responseMap = new HashMap<>();
+        String token = request.getHeader("token");
+        boolean checkValid = AuthService.checkValid(token);
+        if(!checkValid) {
+            responseMap.put("message", "Not authorize!");
+            responseMap.put("data", checkValid);
+            return new ResponseEntity<>(responseMap, HttpStatus.OK);
+        }
+        
         try{
             profile_officer profile_officer = Profile_officerService.findOne(id);
             if(profile_officer.getOfficer_id() == null)
@@ -113,9 +133,17 @@ public class profile_officerController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Object> removeOne(@PathVariable("id") String id)
+    public ResponseEntity<Object> removeOne(@PathVariable("id") String id, HttpServletRequest request)
     {
         Map<String, Object> responseMap = new HashMap<>();
+        String token = request.getHeader("token");
+        boolean checkValid = AuthService.checkValid(token);
+        if(!checkValid) {
+            responseMap.put("message", "Not authorize!");
+            responseMap.put("data", checkValid);
+            return new ResponseEntity<>(responseMap, HttpStatus.OK);
+        }
+        
         Profile_officerService.removeOne(id);
         responseMap.put("message", "Officer successfully deleted!");
         return new ResponseEntity<>(responseMap, HttpStatus.OK);
